+++ src/Stride.CommunityToolkit.DebugShapes/Code/ImmediateDebugRenderSystem.cs
@@ -13,7 +13,7 @@ namespace Stride.CommunityToolkit.DebugShapes.Code;
 /// <summary>
 /// Provides a system for immediate debug rendering of various shapes and lines in the game world.
 /// </summary>
-public partial class ImmediateDebugRenderSystem : GameSystemBase
+public class ImmediateDebugRenderSystem : GameSystemBase
 {
     internal enum DebugRenderableFlags : byte
     {
@@ -30,13 +30,31 @@ public partial class ImmediateDebugRenderSystem : GameSystemBase
     private ImmediateDebugRenderObject? _transparentSolidPrimitiveRenderer;
     private ImmediateDebugRenderObject? _transparentWireframePrimitiveRenderer;
 
+    /// <summary>
+    /// Default color used when a specific color is not provided for a primitive.
+    /// </summary>
     public Color PrimitiveColor { get; set; } = Color.LightGreen;
 
+    /// <summary>
+    /// Maximum number of one-frame primitives stored and rendered per update.
+    /// </summary>
     public int MaxPrimitives { get; set; } = 100;
+
+    /// <summary>
+    /// Maximum number of lifetime-based primitives stored and rendered per update.
+    /// </summary>
     public int MaxPrimitivesWithLifetime { get; set; } = 100;
 
+    /// <summary>
+    /// Render group used for the internal debug render objects.
+    /// </summary>
     public RenderGroup RenderGroup { get; set; }
 
+    /// <summary>
+    /// Creates a new instance of the debug render system.
+    /// </summary>
+    /// <param name="registry">The service registry.</param>
+    /// <param name="renderGroup">The render group to use for the debug render objects.</param>
     public ImmediateDebugRenderSystem(IServiceRegistry registry, RenderGroup renderGroup = RenderGroup.Group31) : base(registry)
     {
         Enabled = true;
@@ -70,6 +88,14 @@ public partial class ImmediateDebugRenderSystem : GameSystemBase
         }
     }
 
+    /// <summary>
+    /// Draws a line between two points.
+    /// </summary>
+    /// <param name="start">Start position.</param>
+    /// <param name="end">End position.</param>
+    /// <param name="color">Line color. Uses <see cref="PrimitiveColor"/> when default.</param>
+    /// <param name="duration">Duration in seconds the line remains visible. 0 draws for a single frame.</param>
+    /// <param name="depthTest">Whether depth testing is enabled.</param>
     public void DrawLine(Vector3 start, Vector3 end, Color color = default, float duration = 0.0f, bool depthTest = true)
     {
         var cmd = new Line { Start = start, End = end, Color = color == default ? PrimitiveColor : color };
@@ -77,6 +103,13 @@ public partial class ImmediateDebugRenderSystem : GameSystemBase
         PushMessage(ref msg);
     }
 
+    /// <summary>
+    /// Draws multiple independent line segments from a vertex list (pairs form segments).
+    /// </summary>
+    /// <param name="vertices">Array of vertices; each consecutive pair defines a line.</param>
+    /// <param name="color">Color for all lines. Uses <see cref="PrimitiveColor"/> when null.</param>
+    /// <param name="duration">Duration in seconds the lines remain visible. 0 draws for a single frame.</param>
+    /// <param name="depthTest">Whether depth testing is enabled.</param>
     public void DrawLines(Vector3[] vertices, Color? color = null, float duration = 0.0f, bool depthTest = true)
     {
         var totalVertexPairs = vertices.Length - (vertices.Length % 2);
@@ -88,17 +121,45 @@ public partial class ImmediateDebugRenderSystem : GameSystemBase
         }
     }
 
+    /// <summary>
+    /// Draws a ray from a starting point along a direction.
+    /// </summary>
+    /// <param name="start">Start position.</param>
+    /// <param name="dir">Direction vector.</param>
+    /// <param name="color">Ray color. Uses <see cref="PrimitiveColor"/> when default.</param>
+    /// <param name="duration">Duration in seconds the ray remains visible. 0 draws for a single frame.</param>
+    /// <param name="depthTest">Whether depth testing is enabled.</param>
     public void DrawRay(Vector3 start, Vector3 dir, Color color = default, float duration = 0.0f, bool depthTest = true)
     {
         DrawLine(start, start + dir, color == default ? PrimitiveColor : color, duration, depthTest);
     }
 
+    /// <summary>
+    /// Draws an arrow composed of a line and a cone tip.
+    /// </summary>
+    /// <param name="from">Arrow start position.</param>
+    /// <param name="to">Arrow end position (tip base).</param>
+    /// <param name="coneHeight">Height of the cone tip.</param>
+    /// <param name="coneRadius">Radius of the cone tip.</param>
+    /// <param name="color">Arrow color. Uses <see cref="PrimitiveColor"/> when default.</param>
+    /// <param name="duration">Duration in seconds the arrow remains visible. 0 draws for a single frame.</param>
+    /// <param name="depthTest">Whether depth testing is enabled.</param>
+    /// <param name="solid">Whether the cone tip is rendered solid or wireframe.</param>
     public void DrawArrow(Vector3 from, Vector3 to, float coneHeight = 0.25f, float coneRadius = 0.125f, Color color = default, float duration = 0.0f, bool depthTest = true, bool solid = false)
     {
         DrawLine(from, to, color, duration, depthTest);
         DrawCone(to, coneHeight, coneRadius, Quaternion.BetweenDirections(new Vector3(0.0f, 1.0f, 0.0f), to - from), color == default ? PrimitiveColor : color, duration, depthTest, solid);
     }
 
+    /// <summary>
+    /// Draws a sphere.
+    /// </summary>
+    /// <param name="position">Sphere center position.</param>
+    /// <param name="radius">Sphere radius.</param>
+    /// <param name="color">Color. Uses <see cref="PrimitiveColor"/> when default.</param>
+    /// <param name="duration">Duration in seconds the sphere remains visible. 0 draws for a single frame.</param>
+    /// <param name="depthTest">Whether depth testing is enabled.</param>
+    /// <param name="solid">Whether rendered solid or wireframe.</param>
     public void DrawSphere(Vector3 position, float radius, Color color = default, float duration = 0.0f, bool depthTest = true, bool solid = false)
     {
         var cmd = new Sphere { Position = position, Radius = radius, Color = color == default ? PrimitiveColor : color };
@@ -107,6 +168,16 @@ public partial class ImmediateDebugRenderSystem : GameSystemBase
         PushMessage(ref msg);
     }
 
+    /// <summary>
+    /// Draws a half-sphere.
+    /// </summary>
+    /// <param name="position">Half-sphere center position.</param>
+    /// <param name="radius">Half-sphere radius.</param>
+    /// <param name="color">Color. Uses <see cref="PrimitiveColor"/> when default.</param>
+    /// <param name="rotation">Rotation applied to the half-sphere.</param>
+    /// <param name="duration">Duration in seconds the half-sphere remains visible. 0 draws for a single frame.</param>
+    /// <param name="depthTest">Whether depth testing is enabled.</param>
+    /// <param name="solid">Whether rendered solid or wireframe.</param>
     public void DrawHalfSphere(Vector3 position, float radius, Color color = default, Quaternion rotation = default, float duration = 0.0f, bool depthTest = true, bool solid = false)
     {
         var cmd = new HalfSphere { Position = position, Radius = radius, Rotation = rotation == default ? Quaternion.Identity : rotation, Color = color == default ? PrimitiveColor : color };
@@ -115,6 +186,16 @@ public partial class ImmediateDebugRenderSystem : GameSystemBase
         PushMessage(ref msg);
     }
 
+    /// <summary>
+    /// Draws axis-aligned bounds represented by a cube.
+    /// </summary>
+    /// <param name="start">Start position of the bounding box.</param>
+    /// <param name="end">End position of the bounding box.</param>
+    /// <param name="rotation">Rotation applied to the box.</param>
+    /// <param name="color">Color. Uses <see cref="PrimitiveColor"/> when default.</param>
+    /// <param name="duration">Duration in seconds the bounds remain visible. 0 draws for a single frame.</param>
+    /// <param name="depthTest">Whether depth testing is enabled.</param>
+    /// <param name="solid">Whether rendered solid or wireframe.</param>
     public void DrawBounds(Vector3 start, Vector3 end, Quaternion rotation = default, Color color = default, float duration = 0.0f, bool depthTest = true, bool solid = false)
     {
         var cmd = new Cube { Start = start + ((end - start) / 2), End = end + ((end - start) / 2), Rotation = rotation == default ? Quaternion.Identity : rotation, Color = color == default ? PrimitiveColor : color };
@@ -148,6 +229,17 @@ public partial class ImmediateDebugRenderSystem : GameSystemBase
         PushMessage(ref msg);
     }
 
+    /// <summary>
+    /// Draws a capsule.
+    /// </summary>
+    /// <param name="position">Capsule center position.</param>
+    /// <param name="height">Capsule height.</param>
+    /// <param name="radius">Capsule radius.</param>
+    /// <param name="rotation">Rotation applied to the capsule.</param>
+    /// <param name="color">Color. Uses <see cref="PrimitiveColor"/> when default.</param>
+    /// <param name="duration">Duration in seconds the capsule remains visible. 0 draws for a single frame.</param>
+    /// <param name="depthTest">Whether depth testing is enabled.</param>
+    /// <param name="solid">Whether rendered solid or wireframe.</param>
     public void DrawCapsule(Vector3 position, float height, float radius, Quaternion rotation = default, Color color = default, float duration = 0.0f, bool depthTest = true, bool solid = false)
     {
         // FIXME: height is divided by two here but can probably be solved more elegantly elsewhere yes
@@ -157,6 +249,17 @@ public partial class ImmediateDebugRenderSystem : GameSystemBase
         PushMessage(ref msg);
     }
 
+    /// <summary>
+    /// Draws a cylinder.
+    /// </summary>
+    /// <param name="position">Cylinder center position.</param>
+    /// <param name="height">Cylinder height.</param>
+    /// <param name="radius">Cylinder radius.</param>
+    /// <param name="rotation">Rotation applied to the cylinder.</param>
+    /// <param name="color">Color. Uses <see cref="PrimitiveColor"/> when default.</param>
+    /// <param name="duration">Duration in seconds the cylinder remains visible. 0 draws for a single frame.</param>
+    /// <param name="depthTest">Whether depth testing is enabled.</param>
+    /// <param name="solid">Whether rendered solid or wireframe.</param>
     public void DrawCylinder(Vector3 position, float height, float radius, Quaternion rotation = default, Color color = default, float duration = 0.0f, bool depthTest = true, bool solid = false)
     {
         var cmd = new Cylinder { Position = position, Height = height, Radius = radius, Rotation = rotation == default ? Quaternion.Identity : rotation, Color = color == default ? PrimitiveColor : color };
@@ -165,6 +268,17 @@ public partial class ImmediateDebugRenderSystem : GameSystemBase
         PushMessage(ref msg);
     }
 
+    /// <summary>
+    /// Draws a cone.
+    /// </summary>
+    /// <param name="position">Cone center position.</param>
+    /// <param name="height">Cone height.</param>
+    /// <param name="radius">Cone base radius.</param>
+    /// <param name="rotation">Rotation applied to the cone.</param>
+    /// <param name="color">Color. Uses <see cref="PrimitiveColor"/> when default.</param>
+    /// <param name="duration">Duration in seconds the cone remains visible. 0 draws for a single frame.</param>
+    /// <param name="depthTest">Whether depth testing is enabled.</param>
+    /// <param name="solid">Whether rendered solid or wireframe.</param>
     public void DrawCone(Vector3 position, float height, float radius, Quaternion rotation = default, Color color = default, float duration = 0.0f, bool depthTest = true, bool solid = false)
     {
         var cmd = new Cone { Position = position, Height = height, Radius = radius, Rotation = rotation == default ? Quaternion.Identity : rotation, Color = color == default ? PrimitiveColor : color };
@@ -173,6 +287,16 @@ public partial class ImmediateDebugRenderSystem : GameSystemBase
         PushMessage(ref msg);
     }
 
+    /// <summary>
+    /// Draws a quad.
+    /// </summary>
+    /// <param name="position">Quad center position.</param>
+    /// <param name="size">Quad dimensions (X = width, Y = height).</param>
+    /// <param name="rotation">Rotation applied to the quad.</param>
+    /// <param name="color">Color. Uses <see cref="PrimitiveColor"/> when default.</param>
+    /// <param name="duration">Duration in seconds the quad remains visible. 0 draws for a single frame.</param>
+    /// <param name="depthTest">Whether depth testing is enabled.</param>
+    /// <param name="solid">Whether rendered solid or wireframe.</param>
     public void DrawQuad(Vector3 position, Vector2 size, Quaternion rotation = default, Color color = default, float duration = 0.0f, bool depthTest = true, bool solid = false)
     {
         var cmd = new Quad { Position = position, Size = size, Rotation = rotation == default ? Quaternion.Identity : rotation, Color = color == default ? PrimitiveColor : color };
@@ -181,6 +305,16 @@ public partial class ImmediateDebugRenderSystem : GameSystemBase
         PushMessage(ref msg);
     }
 
+    /// <summary>
+    /// Draws a circle.
+    /// </summary>
+    /// <param name="position">Circle center position.</param>
+    /// <param name="radius">Circle radius.</param>
+    /// <param name="rotation">Rotation applied to the circle plane.</param>
+    /// <param name="color">Color. Uses <see cref="PrimitiveColor"/> when default.</param>
+    /// <param name="duration">Duration in seconds the circle remains visible. 0 draws for a single frame.</param>
+    /// <param name="depthTest">Whether depth testing is enabled.</param>
+    /// <param name="solid">Whether rendered solid or wireframe.</param>
     public void DrawCircle(Vector3 position, float radius, Quaternion rotation = default, Color color = default, float duration = 0.0f, bool depthTest = true, bool solid = false)
     {
         var cmd = new Circle { Position = position, Radius = radius, Rotation = rotation == default ? Quaternion.Identity : rotation, Color = color == default ? PrimitiveColor : color };
